Add linter config
Since `golangci-lint` has already been pulled into this repository as a
dependency, we should set it up so that it lints more than just the
default, which isn't much.

Add the following presets to our golangci-lint config:
- bugs
- performance
- test

Also exclude the `exhaustruct` linter, since zero-initialization of
our structs during de-serialization is good enough for the semantics we
need.

A list of available linters can be found on golangci-lint's
documentation[^1].

[^1]: https://golangci-lint.run/usage/linters/

Signed-off-by: Andy Sadler <ansadler@redhat.com>