add virtual labels and annotations for access (#147)
We want to return the callers information about why the namespace is
visible to them.

It could be either because a RoleBinding exists for their identity
(user, serviceaccount) or because a RoleBinding exists for one of the
group they participate to.

This change will help the UI to discriminate if visibility on a
namespace is granted because the namespace has been shared with the
group `system:authenticated` or because the user has direct access
(a RoleBinding explicitly targeting the user exists).

For the current implementation, only the first reason is returned to
the user.
That means that, if a user is at the same time granted access to a
Namespace directly and also by sharing the namespace with the
`system:authenticated` group, the user will initially see the virtual
labels/annotations mentioning the direct access. If the direct access
is revoked, then the user will see the group access.

Direct access always has precedence over group access.

The cache is a map where keys are kubernetes subjects (User,
ServiceAccount, or Group) and values are the namespaces the subject
has access to.
With the current implementation when the list of namespaces is built,
subject's main info are copied as virtual labels/annotations in the
namespace.

When a new request comes, the namespace-lister builds the response by
merging the namespaces for each subject matched in the following order:
1. identity of the actor making the call, right now the `name` of the
  user making the call
2. groups of the actor making the call in the order read from the
   request

Only the first occurrence of the namespace is used, so if the user has
direct-access to a namespace it will retrieve the version of the
namespace with labels/annotations set to the user itself. If more groups
are passed, the first one matched will be shown.

Signed-off-by: Francesco Ilario <filario@redhat.com>